refactor: add mail and datasource properties to application-*.yaml files; change structure of the lesson package
