+++ src/main/java/org/example/learningsystem/course/controller/CourseController.java
@@ -8,14 +8,14 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.example.learningsystem.course.dto.CourseRequestDto;
 import org.example.learningsystem.course.dto.CourseResponseDto;
-import org.example.learningsystem.lesson.dto.LessonRequestDto;
-import org.example.learningsystem.lesson.dto.LessonResponseDto;
+import org.example.learningsystem.lesson.common.dto.LessonRequestDto;
+import org.example.learningsystem.lesson.common.dto.LessonResponseDto;
 import org.example.learningsystem.student.dto.StudentResponseDto;
 import org.example.learningsystem.course.mapper.CourseMapper;
-import org.example.learningsystem.lesson.mapper.LessonMapper;
+import org.example.learningsystem.lesson.common.mapper.LessonMapper;
 import org.example.learningsystem.student.mapper.StudentMapper;
 import org.example.learningsystem.course.service.CourseService;
-import org.example.learningsystem.lesson.service.LessonService;
+import org.example.learningsystem.lesson.common.service.LessonService;
 import org.example.learningsystem.student.service.StudentService;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
