+++ src/main/java/org/example/learningsystem/lesson/classroom/dto/ClassroomLessonRequestDto.java
@@ -1,5 +1,6 @@
-package org.example.learningsystem.lesson.dto.videolesson;
+package org.example.learningsystem.lesson.classroom.dto;
 
+import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
@@ -7,21 +8,19 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.SuperBuilder;
-import org.example.learningsystem.lesson.dto.LessonRequestDto;
-import org.hibernate.validator.constraints.URL;
+import org.example.learningsystem.lesson.common.dto.LessonRequestDto;
 
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 @SuperBuilder(toBuilder = true)
-public class VideoLessonRequestDto extends LessonRequestDto {
+public class ClassroomLessonRequestDto extends LessonRequestDto {
 
     @NotBlank
-    @Size(max = 256)
-    @URL
-    private String url;
-
     @Size(max = 100)
-    private String platform;
+    private String location;
+
+    @Min(1)
+    private Integer capacity;
 }
