+++ src/main/java/org/example/learningsystem/lesson/video/dto/VideoLessonResponseDto.java
@@ -1,9 +1,9 @@
-package org.example.learningsystem.lesson.dto.videolesson;
+package org.example.learningsystem.lesson.video.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 import lombok.Setter;
-import org.example.learningsystem.lesson.dto.LessonResponseDto;
+import org.example.learningsystem.lesson.common.dto.LessonResponseDto;
 
 @Schema
 @Getter
