+++ src/test/integration/org/example/learningsystem/lesson/controller/LessonControllerIT.java
@@ -5,9 +5,9 @@ import org.example.learningsystem.AbstractCommonIT;
 import org.example.learningsystem.core.exception.model.EntityNotFoundException;
 import org.example.learningsystem.course.model.Course;
 import org.example.learningsystem.course.service.CourseService;
-import org.example.learningsystem.lesson.model.ClassroomLesson;
-import org.example.learningsystem.lesson.model.VideoLesson;
-import org.example.learningsystem.lesson.service.LessonService;
+import org.example.learningsystem.lesson.classroom.model.ClassroomLesson;
+import org.example.learningsystem.lesson.video.model.VideoLesson;
+import org.example.learningsystem.lesson.common.service.LessonService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
