+++ src/test/resources/application-test.yaml
@@ -5,3 +5,13 @@ spring:
   datasource:
     username: test
     password: test
+
+mail:
+  from: hello@demomailtrap.co
+  user: smtp@mailtrap.io
+  password:
+  host: live.smtp.mailtrap.io
+  port: 587
+  protocol: smtp
+  auth: true
+  start-tls-enable: true
